cinemaProject/backend/requirements.txt
asgiref==3.9.2
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
sqlparse==0.5.3
cryptography==44.0.0
django-encrypted-model-fields==0.6.5
